requirements.txt
click>=6.7
Flask>=1.0.0
Flask-HTTPAuth>=3.3.0
Flask-Migrate>=2.0.4
Flask-SQLAlchemy>=2.3.2
flask-restful>=0.3
itsdangerous>=0.24
Jinja2>=2.10
PyMySQL>=0.9.2
pluggy>=0.6.0
pytz>=2018.4
SQLAlchemy-Utils>=0.33.3
Werkzeug>=0.14.1
pyecharts>=1.0.0
-e .
